competitive-llms/requirements.txt
accelerate==0.20.3
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
blinker==1.4
cachetools==5.3.1
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.1.0
Click==7.0
cmake==3.26.4
cohere==4.11.2
colorama==0.4.3
cryptography==2.8
datasets==2.13.0
dill==0.3.6
diskcache==5.6.1
distro==1.4.0
einops==0.6.1
entrypoints==0.3
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
gptcache==0.1.32
guidance==0.0.63
httplib2==0.14.0
huggingface-hub==0.15.1
idna==2.8
importlib-metadata==6.6.0
Jinja2==3.1.2
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lit==16.0.6
macaroonbakery==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
msal==1.22.0
msgpack==0.6.2
multidict==6.0.4
multiprocess==0.70.14
nest-asyncio==1.5.6
netifaces==0.10.4
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.0
olefile==0.46
openai==0.27.8
packaging==23.1
pandas==2.0.2
peft==0.3.0
pexpect==4.6.0
Pillow==7.0.0
platformdirs==3.6.0
protobuf==3.20.1
psutil==5.9.5
pyarrow==12.0.1
pycairo==1.16.2
Pygments==2.15.1
pygtrie==2.5.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynvim==0.4.1
pyparsing==3.0.9
pyRFC3339==1.1
python-dateutil==2.8.2
pytz==2023.3
pyxdg==0.26
PyYAML==5.3.1
regex==2023.6.3
reportlab==3.5.34
requests>=2.26.0
requests-unixsocket==0.2.0
rich==13.4.2
safetensors==0.3.1
SecretStorage==2.3.1
sentencepiece==0.1.99
simplejson==3.16.0
six==1.14.0
ssh-import-id==5.10
sympy==1.12
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.30.2
triton==2.0.0
typing-extensions==4.6.3
tzdata==2023.3
urllib3==1.25.8
wadllib==1.3.3
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
